// app/shared/types/enums.ts
export enum TaskPriority {
  URGENT = "urgent",
  HIGH = "high",
  MEDIUM = "medium",
  LOW = "low",
  NONE = "none",
}

export enum TaskStatus {
  NOT_STARTED = "not_started",
  IN_PROGRESS = "in_progress",
  COMPLETED = "completed",
}

export enum CustomFieldType {
  TEXT = "text",
  NUMBER = "number",
  CHECKBOX = "checkbox",
}

export enum ViewMode {
  TABLE = "table",
  KANBAN = "kanban",
}

export enum SortDirection {
  ASC = "asc",
  DESC = "desc",
}

export enum FieldType {
  TEXT = "text",
  RADIO = "radio",
  CHECKBOX = "checkbox",
  NUMBER = "number",
  SELECT = "select",
}

export enum Size {
  XS = "xs",
  SM = "sm",
  MD = "md",
}